+++ crates/optimism/rpc/src/lib.rs
@@ -11,6 +11,7 @@
 pub mod engine;
 pub mod error;
 pub mod eth;
+pub mod historical;
 pub mod miner;
 pub mod sequencer;
 pub mod witness;
